jhs cli 添加转存集换社心愿单功能

+++ pkg/sdk/jihuanshe/services/wishes/models/wish_list_create_resp.go
@@ -4,5 +4,5 @@ type WishListCreateResp struct {
 	GameKey    string `json:"game_key"`
 	GameSubKey string `json:"game_sub_key"`
 	Name       string `json:"name"`
-	WishListID int64  `json:"wish_list_id"`
+	WishListID int    `json:"wish_list_id"`
 }
